requirements.txt
pdfplumber
pytesseract
pdf2image
python-dotenv
PyYAML
# numpy
# pytest
watchdog 